refactor(water-cost-calculations): optimize loop, deduplicate parseAsUtc, and fix timezone comparisons
